Supercharged View Template handling
FOF can now ship with default (common) View Templates in its own
`ViewTemplates` folder. You can override them per template in its
`<templateFolder>/html/lib_fof40` folder.

Automatic suffixes on PHP and Blade view templates depending on the
Joomla! version and renderer being used, parsed from most to least
specific. For example:

* default.j39.fef.php
* default.j39.php
* default.j3.fef.php
* default.j3.php
* default.fef.php
* default.php

This does not override the path priority which remains: template
overrides, specified side of the application, opposite side of the
application.

If a less specific view template is found in a higher priority path it
will be used instead. The idea being that a template override is always
most specific as it's done for a specific site which runs a specific
Joomla! version and for a specific component whose renderer you already
know.