Refactored Input
Nothing to change since we're extending Joomla's
Input class and we're bound by its definition of
public methods. Remember that until PHP 8 we cannot
have covariance or contravariance, meaning we cannot
go from implied to explicit types in a descendant
class even when that change would actually be type
safe.