Renderers can now load CSS and JS files before the onBefore* view events are triggered
This means that any JS or CSS files loaded by the view events
will be loaded _after_ the framework files loaded by the
Renderer. This makes far more sense as an architecture and can
lead to the decrease of the stuff we were loading in the
Dispatcher of each component.