Support for Joomla 4-style view template folders
Joomla 4 stored the view templates in the `tmpl` folder under the
component area's root in the same way we use the `ViewTemplates` folder.

Joomla 4's template manager _assumes_ this is where view templates live.
So, we need to adjust FOF to allow Joomla's template manager figure its
crap out.

I have to say I'm flattered that Joomla tried to copy the concepts of
FOF 3. I am a bit miffed they failed to understand the full reasoning
behind them and botched the transition. Well, it is what it is.